+++ src/assets/formatted/AdminPencil.tsx
@@ -1,0 +1,39 @@
+import { forwardRef, memo } from 'react';
+import { kebabCase } from 'change-case';
+
+interface AdminPencilProps extends React.HTMLAttributes<SVGSVGElement> {
+  color?: string;
+  height?: number | string;
+  width?: number | string;
+  title?: string;
+  titleId?: string;
+}
+
+const AdminPencil = memo(
+  forwardRef<SVGSVGElement, AdminPencilProps>((props, ref) => {
+    const { color, height, width, title, titleId: propsTitleId } = props;
+    const titleId = propsTitleId || kebabCase(title || '');
+
+    return (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        fill="none"
+        viewBox="0 0 24 24"
+        height={height}
+        width={width}
+        role="img"
+        ref={ref}
+        aria-labelledby={titleId}
+        {...props}
+      >
+        {title ? <title id={titleId}>{title}</title> : null}
+        <path
+          fill={color || '#fff'}
+          d="M3 17.25V21h3.75L17.81 9.94l-3.75-3.75zM20.71 7.04a.996.996 0 0 0 0-1.41l-2.34-2.34a.996.996 0 0 0-1.41 0l-1.83 1.83 3.75 3.75z"
+        />
+      </svg>
+    );
+  }),
+);
+
+export default AdminPencil;