+++ src/components/Input/Input.stories.tsx
@@ -41,6 +41,17 @@ const argTypes = {
       type: 'text',
     },
   },
+  inputAdornment: {
+    control: {
+      type: 'text',
+    },
+  },
+  adornmentPosition: {
+    options: ['start', 'end'],
+    control: {
+      type: 'select',
+    },
+  },
   labelText: {
     control: {
       type: 'text',
@@ -238,6 +249,22 @@ CustomLabel.args = {
 };
 
 CustomLabel.argTypes = {};
+
+export const InputWithAdornment = ({ playgroundWidth, ...args }: StoryProps) => (
+  <div style={{ width: playgroundWidth, margin: '1rem' }}>
+    <Input key={args.defaultValue as string} {...args} id="Input-1" />
+  </div>
+);
+
+InputWithAdornment.args = {
+  labelText: 'Input with Adornment',
+  playgroundWidth: 300,
+  size: 'md',
+  inputAdornment: <Text variant={TextVariants.string2}>$</Text>,
+  adornmentPosition: 'start',
+};
+
+InputWithAdornment.argTypes = argTypes;
 
 export const Playground = ({ playgroundWidth, ...args }: StoryProps) => {
   const [loading, setLoading] = useState(false);